refactor(DEVING-47): 모임 상세 리펙토링
