+++ src/app/meeting/components/UserInfo.tsx
@@ -13,9 +13,18 @@ const UserInfo = ({ meetingId }: { meetingId: number }) => {
     return <SkeletonUserInfo />;
   }
 
-  console.log('userinfo: ', data);
+  // const data = {
+  //   name: 'coding1234',
+  //   profilePic:
+  //     'https://deving-bucket.s3.ap-northeast-2.amazonaws.com/profile_img.png',
+  //   email: 'qwe@qwe.qwe',
+  //   intro: 'Nice to meet you',
+  //   phone: null,
+  //   skillArray: [],
+  // };
+
   return (
-    <div className="flex flex-col gap-[16px] p-[16px] md:px-[32px]">
+    <div className="flex flex-col gap-[16px]">
       <div className="flex items-center gap-[16px]">
         <div className="relative h-[80px] w-[80px] shrink-0">
           <Image
