+++ src/app/meeting/components/skeletons/SkeletonMeetingTotalInfo.tsx
@@ -3,15 +3,18 @@ import SkeletonBasic from '../../../../components/ui/SkeletonBasic';
 const SkeletonMeetingTotalInfo = () => {
   return (
     <>
-      <div className="mb-[32px] hidden h-[240px] w-full flex-shrink-0 flex-row gap-[40px] p-[32px] lg:flex">
+      {/* 테스크탑 */}
+      <div className=" hidden w-full flex-row gap-[40px] lg:flex">
         <SkeletonBasic className="h-[208px] flex-1" />
         <SkeletonBasic className="h-[208px] w-[318px]" />
       </div>
-      <div className="hidden w-full flex-shrink-0 flex-col gap-[40px] p-8 md:flex lg:hidden">
-        <SkeletonBasic className="h-[252px] w-full" />
+      {/* 태블릿 */}
+      <div className="hidden w-full flex-col gap-[32px] md:flex lg:hidden">
+        <SkeletonBasic className="h-[210px] w-full" />
         <SkeletonBasic className="h-[133px] w-full" />
       </div>
-      <div className="flex w-full flex-shrink-0 flex-col gap-[16px] p-4 md:hidden">
+      {/* 모바일 */}
+      <div className="mb-[16px] flex w-full flex-col gap-[16px] md:hidden">
         <SkeletonBasic className="h-[252px] w-full" />
         <SkeletonBasic className="h-[86px] w-full" />
         <SkeletonBasic className="h-[133px] w-full" />
