+++ src/app/meeting/study/[id]/page.tsx
@@ -4,14 +4,13 @@ import { QueryClient } from '@tanstack/react-query';
 import { ArrowLeft } from 'lucide-react';
 
 import CardWarpper from '../../components/CardWrapper';
+import ContentLabel from '../../components/ContextLabel';
 import MeetingInfo from '../../components/MeetingInfo';
 import ReviewInput from '../../components/ReviewInput';
 import ReviewList from '../../components/ReviewList';
 import UserInfo from '../../components/UserInfo';
 
 export default function page({ params }: { params: { id: string } }) {
-  console.log('params: ', params);
-
   const queryClient = new QueryClient();
   const meetingId = parseInt(params.id);
 
@@ -26,16 +25,11 @@ export default function page({ params }: { params: { id: string } }) {
       <div className="mt-[40px]">
         <ArrowLeft className="h-[24px] w-[24px] text-Cgray700" />
       </div>
-      <div className="flex flex-col gap-[48px]">
+      <div className="flex flex-col gap-[48px] p-[16px] md:px-[48px]">
         <CardWarpper meetingId={meetingId} />
         <UserInfo meetingId={meetingId} />
         <MeetingInfo meetingId={meetingId} />
-      </div>
-      <div className="flex flex-col gap-[48px] p-[16px] md:px-[48px]">
-        <div className="mb-[-24px] flex items-center gap-[8px]">
-          <div className="h-[14px] w-[2px] bg-Cgray700" />
-          <h3 className="typo-head3 text-Cgray700">리뷰</h3>
-        </div>
+        <ContentLabel>리뷰</ContentLabel>
         <ReviewAvgCard meetingId={meetingId} />
         <ReviewInput meetingId={meetingId} />
         <ReviewList meetingId={meetingId} />
